Add a --nolibpython option to valabind-cc.
From Debian's Python Policy, Chapter 2.1:

Some distributions link extensions to libpython, but this is
not the case in Debian as symbols might as well be resolved
by /usr/bin/pythonX.Y which is not linked to libpython.

This patch adds a new option to valabind-cc to strip the -lpythonX.X from
the output of pythonX.Y-config --ldflags.  It also makes the matching of
arguments more strict to prevent '-n' being matched with --nolibpython.